requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bokeh==3.4.0
Brotli==1.0.9
bs4==0.0.2
cachelib==0.9.0
cachetools==5.3.3
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.5
contourpy==1.2.0
curlify==2.2.1
dash==2.6.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-extensions==0.1.6
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
dataclasses-json==0.5.14
EditorConfig==0.12.3
et-xmlfile==1.1.0
facebook-sdk==3.1.0
facebook_business==19.0.2
Flask==2.1.0
Flask-Caching==2.0.0
Flask-Compress==1.12
frozenlist==1.4.1
future==0.18.2
google-ads==23.1.0
google-api-core==2.18.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==2.8
itsdangerous==2.1.2
Jinja2==3.1.2
jsbeautifier==1.14.6
lxml==5.2.1
MarkupSafe==2.1.1
marshmallow==3.21.1
more-itertools==8.14.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.23.3
oauthlib==3.2.2
openpyxl==3.0.10
packaging==24.0
pandas==1.4.4
pillow==10.2.0
plotly==5.10.0
product==0.1.1.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==23.12.11
python-dateutil==2.8.2
python-dotenv==1.0.1
python-facebook-api==0.19.0
pytz==2022.2.1
PyYAML==6.0.1
requests==2.22.0
requests-oauthlib==1.3.1
retrying==1.3.3
rsa==4.9
six==1.16.0
soupsieve==2.5
tenacity==8.0.1
tornado==6.4
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==1.25.11
Werkzeug==2.2.2
xyzservices==2023.10.1
yarl==1.9.4
glob2